Add Remappers for Toggle Switches
Add three new remappers to support toggle switches.

Add a remapper SwitchToButtons that creates a different momentary button press when a toggle switch is moved to a new possiton.  This is used to allow a two possition toggle switch to map to game controls that have a seperate button for each in game switch position.  (i.e. "Landing Gear Up" and "Landing Gear Down")

Add a remapper SwitchToToggle that creates a momentary press of the same button each time a toggle switch is moved to a new possition.  This is to allow a two position toggle switch to map to game controls that only have a single key to toggle a switch. (i.e. "Toggle Landing Lights").  This will also only create one button press if the switch is flipped back and forth quickly so that if game and physical switch are out of sync (one on and other is off) a quick flip back and forth will put them back in sync.

Add a remapper "Switch3WayToButtons" that creates a different momentary button press when an ON-OFF-ON toggle switch is moved to a each possiton.  This is used to allow a three possition ON-OFF-ON toggle switch connected to two buttons to map to game controls that have a seperate button for each of three in game switch position.  (i.e. "Headlights Off" and "Headlights On" and "Headlight High Beam")